[REF] web,web_editor: move legacy utils attachComponent
This utils allows to spawn an owl Component inside a legacy widget.
We're at the end of the owl conversion, and this legacy util is
only used in web_editor (which has already been converted, but both
implementations still co-exist, see html_editor addon). We thus
move the util to web_editor, s.t. it will be removed alongside the
addon.
