+++ addons/web_editor/static/src/js/core/owl_utils.js
