+++ addons/web_editor/static/src/js/editor/snippets.options.js
@@ -1,6 +1,6 @@
 /** @odoo-module **/
 
-import { attachComponent } from "@web/legacy/utils";
+import { attachComponent } from "@web_editor/js/core/owl_utils";
 import { MediaDialog } from "@web_editor/components/media_dialog/media_dialog";
 import { ConfirmationDialog } from "@web/core/confirmation_dialog/confirmation_dialog";
 import { throttleForAnimation, debounce } from "@web/core/utils/timing";
